requirements.txt
moviepy==1.0.3
google_auth_oauthlib==0.3.0
opencv_python==4.5.3.56
requests==2.21.0
hurry.filesize==0.9
watson_developer_cloud==2.8.0
pymediawiki==0.7.0
numpy==1.21.2
nltk>=3.6.5
Algorithmia==1.13.1
google_api_python_client==2.26.1
hurry==1.1
Pillow==8.4.0
mysql-connector-python==8.0.28